CUDA GPU implementations and tests

+++ package/miga/types.pxd
@@ -4,4 +4,4 @@ from libc.stdint cimport int32_t
 
 ctypedef int32_t seq_t;
 ctypedef int32_t index_t;
-ctypedef double data_t;
+ctypedef float data_t;
